+++ packages/core/src/types/curi.ts
@@ -49,6 +49,7 @@ export interface NavigationDetails {
   state?: any;
   method?: NavType;
   cancelled?: () => void;
+  finished?: () => void;
 }
 
 export interface CuriRouter {